// types/server/contracts.d.ts
  interface Profile {
    name: string;
    overview: string;
    un_sdg?: number;
    tier?: number;
    logo?: string;
    image?: string;
    url?: string;
    registration_number?: string;
    country_city_origin?: string;
    contact_email?: string;
    social_media_urls: {
      facebook?: string;
      linkedin?: string;
      twitter?: string;
    };
    number_of_employees?: number;
    average_annual_budget?: string;
    annual_revenue?: string;
    charity_navigator_rating?: string;
  }
// ...
  type CW3Config = {
    group_addr: string; //"terra123abc.."
    threshold: {
      absolute_percentage: {
        percentage: string; //"0.5"
      };
    };
    max_voting_period: {
      height: number; //1000
    };
    isPlacholder?: true;
    //...future needed attr
  };

  type AdminVoteInfo = {
    voter: string; //"terra123abc.."
    vote: Vote;
    weight: number; //1
  };

  type CWContracts = "apTeam" | { cw3?: string; cw4?: string };
// ...
  type EndowmentStatus = {
    Inactive: 0;
    Approved: 1;
    Frozen: 2;
    Closed: 3;
  };

  type EndowmentType = "charity";
  type EndowmentStatusNum = EndowmentStatus[keyof EndowmentStatus];
  type EndowmentStatusStrNum = `${EndowmentStatusNum}`;

  type EndowmentTier = "Level1" | "Level2" | "Level3";
  type EndowmentEntry = {
    address: string;
    status: keyof EndowmentStatus;
    name: string;
    owner?: string;
    tier?: EndowmentTier;
    endow_type?: EndowmentType;
    un_sdg?: number;
  };

  type CategorizedEndowments = {
    [index: number]: EndowmentEntry[];
  };